Refactor note list and tag management for improved UI
This commit refactors the note list and tag management components to improve the user interface. It moves the "Add Note" button into the NoteList component and restyles the TagManagement component to include an "Add Tag" button. It also removes the registration error message from the NoteList component and adjusts the styling of various elements for a cleaner look.